@@ -17,42 +17,51 @@
   ],
   "minimum-stability": "stable",
   "require": {
-    "php": ">=5.6.4",
+    "php": "^7.1.3",
     "asm89/stack-cors": "^1.2",
     "chumper/zipper": "^1.0",
     "doctrine/inflector": "~1.0",
     "dompdf/dompdf": "^0.8.2",
-    "filp/whoops": "^2.1",
-    "illuminate/auth": "^5.5",
-    "illuminate/bus": "^5.5",
-    "illuminate/cache": "^5.5",
-    "illuminate/config": "^5.5",
-    "illuminate/container": "^5.5",
-    "illuminate/contracts": "^5.5",
-    "illuminate/cookie": "^5.5",
-    "illuminate/database": "^5.5",
-    "illuminate/encryption": "^5.5",
-    "illuminate/events": "^5.5",
-    "illuminate/hashing": "^5.5",
-    "illuminate/log": "^5.5",
-    "illuminate/mail": "^5.5",
-    "illuminate/notifications": "^5.5",
-    "illuminate/pagination": "^5.5",
-    "illuminate/queue": "^5.5",
-    "illuminate/routing": "^5.5",
-    "illuminate/support": "^5.5",
-    "illuminate/validation": "^5.5",
-    "illuminate/view": "^5.5",
+    "illuminate/auth": "^5.6",
+    "illuminate/bus": "^5.6",
+    "illuminate/cache": "^5.6",
+    "illuminate/config": "^5.6",
+    "illuminate/console": "^5.6",
+    "illuminate/container": "^5.6",
+    "illuminate/contracts": "^5.6",
+    "illuminate/cookie": "^5.6",
+    "illuminate/database": "^5.6",
+    "illuminate/encryption": "^5.6",
+    "illuminate/events": "^5.6",
+    "illuminate/filesystem": "^5.6",
+    "illuminate/hashing": "^5.6",
+    "illuminate/log": "^5.6",
+    "illuminate/mail": "^5.6",
+    "illuminate/notifications": "^5.6",
+    "illuminate/pagination": "^5.6",
+    "illuminate/queue": "^5.6",
+    "illuminate/redis": "^5.6",
+    "illuminate/routing": "^5.6",
+    "illuminate/support": "^5.6",
+    "illuminate/validation": "^5.6",
+    "illuminate/view": "^5.6",
     "mtdowling/cron-expression": "^1.2",
     "nesbot/carbon": "^1.22",
     "predis/predis": "^1.1",
-    "pusher/pusher-php-server": "~3.0",
-    "robmorgan/phinx": "^0.8.0",
+    "robmorgan/phinx": "^0.9.2",
     "swiftmailer/swiftmailer": "^6.0",
+    "symfony/console": "^4.0",
     "symfony/polyfill-mbstring": "^1.7",
     "symfony/var-dumper": "^3.0",
     "vlucas/phpdotenv": "^2.4",
     "wikimedia/composer-merge-plugin": "^1.4"
+  },
+  "require-dev": {
+    "filp/whoops": "^2.0",
+    "fzaninotto/faker": "^1.4",
+    "mockery/mockery": "^1.0",
+    "nunomaduro/collision": "^2.0",
+    "phpunit/phpunit": "^7.0"
   },
   "autoload": {
     "classmap": [
