@@ -7,47 +7,51 @@
     3 => 'Illuminate\\Bus\\BusServiceProvider',
     4 => 'Illuminate\\Cache\\CacheServiceProvider',
     5 => 'Illuminate\\Cookie\\CookieServiceProvider',
-    6 => 'Illuminate\\Filesystem\\FilesystemServiceProvider',
-    7 => 'Illuminate\\Hashing\\HashServiceProvider',
-    8 => 'Illuminate\\Log\\LogServiceProvider',
+    6 => 'Illuminate\\Encryption\\EncryptionServiceProvider',
+    7 => 'Illuminate\\Filesystem\\FilesystemServiceProvider',
+    8 => 'Illuminate\\Hashing\\HashServiceProvider',
     9 => 'Illuminate\\Mail\\MailServiceProvider',
     10 => 'Illuminate\\Notifications\\NotificationServiceProvider',
     11 => 'Illuminate\\Pagination\\PaginationServiceProvider',
     12 => 'Illuminate\\Pipeline\\PipelineServiceProvider',
     13 => 'Illuminate\\Queue\\QueueServiceProvider',
-    14 => 'Illuminate\\Session\\SessionServiceProvider',
-    15 => 'Illuminate\\Validation\\ValidationServiceProvider',
-    16 => 'Pluma\\Providers\\ViewServiceProvider',
-    17 => 'Pluma\\Providers\\ApplicationServiceProvider',
-    18 => 'Pluma\\Providers\\DatabaseServiceProvider',
-    19 => 'Pluma\\Providers\\EncryptionServiceProvider',
-    20 => 'Pluma\\Providers\\EventServiceProvider',
-    21 => 'Pluma\\Providers\\TranslationServiceProvider',
-    22 => 'Pluma\\Providers\\ModuleServiceProvider',
-    23 => 'Pluma\\Providers\\FormRequestServiceProvider',
-    24 => 'Pluma\\Providers\\CorsServiceProvider',
-    25 => 'Pluma\\Support\\Installation\\Providers\\InstallationServiceProvider',
-    26 => 'Pluma\\Support\\Broadcast\\BroadcastServiceProvider',
-    27 => 'Blacksmith\\Providers\\ConsoleSupportServiceProvider',
+    14 => 'Illuminate\\Redis\\RedisServiceProvider',
+    15 => 'Illuminate\\Session\\SessionServiceProvider',
+    16 => 'Illuminate\\Validation\\ValidationServiceProvider',
+    17 => 'Pluma\\Providers\\ViewServiceProvider',
+    18 => 'Pluma\\Providers\\ApplicationServiceProvider',
+    19 => 'Pluma\\Providers\\DatabaseServiceProvider',
+    20 => 'Pluma\\Providers\\FilesystemServiceProvider',
+    21 => 'Pluma\\Providers\\EncryptionServiceProvider',
+    22 => 'Pluma\\Providers\\EventServiceProvider',
+    23 => 'Pluma\\Providers\\TranslationServiceProvider',
+    24 => 'Pluma\\Providers\\ModuleServiceProvider',
+    25 => 'Pluma\\Providers\\FormRequestServiceProvider',
+    26 => 'Pluma\\Providers\\CorsServiceProvider',
+    27 => 'Pluma\\Support\\Installation\\Providers\\InstallationServiceProvider',
+    28 => 'Pluma\\Support\\Broadcast\\BroadcastServiceProvider',
+    29 => 'Blacksmith\\Providers\\ConsoleSupportServiceProvider',
   ),
   'eager' => 
   array (
     0 => 'Illuminate\\Auth\\AuthServiceProvider',
     1 => 'Illuminate\\Cookie\\CookieServiceProvider',
-    2 => 'Illuminate\\Filesystem\\FilesystemServiceProvider',
-    3 => 'Illuminate\\Log\\LogServiceProvider',
+    2 => 'Illuminate\\Encryption\\EncryptionServiceProvider',
+    3 => 'Illuminate\\Filesystem\\FilesystemServiceProvider',
     4 => 'Illuminate\\Notifications\\NotificationServiceProvider',
     5 => 'Illuminate\\Pagination\\PaginationServiceProvider',
     6 => 'Illuminate\\Session\\SessionServiceProvider',
     7 => 'Pluma\\Providers\\ViewServiceProvider',
-    8 => 'Pluma\\Providers\\DatabaseServiceProvider',
-    9 => 'Pluma\\Providers\\EncryptionServiceProvider',
-    10 => 'Pluma\\Providers\\EventServiceProvider',
-    11 => 'Pluma\\Providers\\ModuleServiceProvider',
-    12 => 'Pluma\\Providers\\FormRequestServiceProvider',
-    13 => 'Pluma\\Providers\\CorsServiceProvider',
-    14 => 'Pluma\\Support\\Installation\\Providers\\InstallationServiceProvider',
-    15 => 'Pluma\\Support\\Broadcast\\BroadcastServiceProvider',
+    8 => 'Pluma\\Providers\\ApplicationServiceProvider',
+    9 => 'Pluma\\Providers\\DatabaseServiceProvider',
+    10 => 'Pluma\\Providers\\FilesystemServiceProvider',
+    11 => 'Pluma\\Providers\\EncryptionServiceProvider',
+    12 => 'Pluma\\Providers\\EventServiceProvider',
+    13 => 'Pluma\\Providers\\ModuleServiceProvider',
+    14 => 'Pluma\\Providers\\FormRequestServiceProvider',
+    15 => 'Pluma\\Providers\\CorsServiceProvider',
+    16 => 'Pluma\\Support\\Installation\\Providers\\InstallationServiceProvider',
+    17 => 'Pluma\\Support\\Broadcast\\BroadcastServiceProvider',
   ),
   'deferred' => 
   array (
@@ -63,6 +67,7 @@
     'cache.store' => 'Illuminate\\Cache\\CacheServiceProvider',
     'memcached.connector' => 'Illuminate\\Cache\\CacheServiceProvider',
     'hash' => 'Illuminate\\Hashing\\HashServiceProvider',
+    'hash.driver' => 'Illuminate\\Hashing\\HashServiceProvider',
     'mailer' => 'Illuminate\\Mail\\MailServiceProvider',
     'swift.mailer' => 'Illuminate\\Mail\\MailServiceProvider',
     'swift.transport' => 'Illuminate\\Mail\\MailServiceProvider',
@@ -73,6 +78,8 @@
     'queue.listener' => 'Illuminate\\Queue\\QueueServiceProvider',
     'queue.failer' => 'Illuminate\\Queue\\QueueServiceProvider',
     'queue.connection' => 'Illuminate\\Queue\\QueueServiceProvider',
+    'redis' => 'Illuminate\\Redis\\RedisServiceProvider',
+    'redis.connection' => 'Illuminate\\Redis\\RedisServiceProvider',
     'validator' => 'Illuminate\\Validation\\ValidationServiceProvider',
     'validation.presence' => 'Illuminate\\Validation\\ValidationServiceProvider',
     'translator' => 'Pluma\\Providers\\TranslationServiceProvider',
@@ -104,10 +111,10 @@
     'Illuminate\\Queue\\QueueServiceProvider' => 
     array (
     ),
-    'Illuminate\\Validation\\ValidationServiceProvider' => 
+    'Illuminate\\Redis\\RedisServiceProvider' => 
     array (
     ),
-    'Pluma\\Providers\\ApplicationServiceProvider' => 
+    'Illuminate\\Validation\\ValidationServiceProvider' => 
     array (
     ),
     'Pluma\\Providers\\TranslationServiceProvider' => 
