@@ -14,7 +14,7 @@ return [
      *
      */
 
-    'default' => env('QUEUE_DRIVER', 'database'),
+    'default' => env('QUEUE_DRIVER', 'sync'),
 
     /**
      *--------------------------------------------------------------------------
