@@ -2,18 +2,14 @@
 
 namespace Pluma\Providers;
 
+use Illuminate\Http\Request;
 use Illuminate\Support\AggregateServiceProvider;
+use Illuminate\Support\Facades\URL;
+use Illuminate\Support\Str;
 use Pluma\Providers\QueueServiceProvider;
 
 class ApplicationServiceProvider extends AggregateServiceProvider
 {
-    /**
-     * Indicates if loading of the provider is deferred.
-     *
-     * @var bool
-     */
-    protected $defer = true;
-
     /**
      * The provider class names.
      *
@@ -21,6 +17,46 @@ class ApplicationServiceProvider extends AggregateServiceProvider
      */
     protected $providers = [
         FormRequestServiceProvider::class,
-        // QueueServiceProvider::class,
     ];
+
+    /**
+     * Register the service provider.
+     *
+     * @return void
+     */
+    public function register()
+    {
+        parent::register();
+
+        $this->registerRequestValidation();
+        $this->registerRequestSignatureValidation();
+    }
+
+    /**
+     * Register the "validate" macro on the request.
+     *
+     * @return void
+     */
+    public function registerRequestValidation()
+    {
+        Request::macro('validate', function (array $rules, ...$params) {
+            validator()->validate($this->all(), $rules, ...$params);
+
+            return $this->only(collect($rules)->keys()->map(function ($rule) {
+                return Str::contains($rule, '.') ? explode('.', $rule)[0] : $rule;
+            })->unique()->toArray());
+        });
+    }
+
+    /**
+     * Register the "hasValidSignature" macro on the request.
+     *
+     * @return void
+     */
+    public function registerRequestSignatureValidation()
+    {
+        Request::macro('hasValidSignature', function () {
+            return URL::hasValidSignature($this);
+        });
+    }
 }
