@@ -100,7 +100,7 @@
                     resource: {
                        item: {!! json_encode(old() ?? []) !!},
                        errors: {!! json_encode($errors->getMessages()) !!},
-                   },
+                    },
                 };
             },
         });
