@@ -14,6 +14,6 @@ trait BelongsToManyCourses
     public function courses()
     {
         return $this->belongsToMany(Course::class)
-                    ->withPivot('dropped_at', 'enrolled_at');
+                    ->withPivot('dropped_at', 'status', 'enrolled_at');
     }
 }
